+++ widgets/EntryParticipants.php
@@ -59,8 +59,15 @@ class EntryParticipants extends Widget
         }
 
         $participantSate = $calendarEntry->getParticipationStatus(Yii::$app->user->identity);
-        return Button::info($label)->sm()
-                     ->icon($participantSate === $state ? 'fa-check' : null)
-                     ->action('calendar.respond', Url::toEntryRespond($calendarEntry, $state));
+
+        $button = Button::info($label)->sm()
+            ->icon($participantSate === $state ? 'fa-check' : null);
+        if ($calendarEntry->isPast()) {
+            $button->cssClass('disabled')->loader(false);
+        } else {
+            $button->action('calendar.respond', Url::toEntryRespond($calendarEntry, $state));
+        }
+
+        return $button;
     }
 }