+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog
 - Enh #187: Possibility to disable participation activities in the mail summary
 - Enh #5274: Deprecate CompatModuleManager
 - Fix #254: Fix wall stream entry icons color
+- Fix: In the calendar view, "Day" button is not translated with Humhub's messages
 
 1.1.10  (April 14, 2021)
 ----------------------