+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog
 - Enh #5274: Deprecate CompatModuleManager
 - Fix #254: Fix wall stream entry icons color
 - Fix #260: In the calendar view, "Day" button is not translated with Humhub's messages
+- Enh #265: Disable to change participant status when event is over
 
 1.1.10  (April 14, 2021)
 ----------------------